requirements.txt
PyYAML>=6.0.0
ruamel.yaml>=0.18.0
numpy==1.26.4
pandas>=2.2.0
scipy>=1.13.0
matplotlib>=3.9.0
torch==2.2.2
tqdm>=4.66.0
hmmlearn>=0.3.0
opencv-python-headless>=4.9.0.0
umap-learn>=0.5.0
h5py>=3.11.0
pydantic==2.7.4
imageio==2.34.1
imageio-ffmpeg==0.5.1
